+++ src/main/java/gitbal/backend/api/auth/service/AuthService.java
@@ -2,7 +2,7 @@ package gitbal.backend.api.auth.service;
 
 import gitbal.backend.api.auth.dto.JoinRequestDto;
 import gitbal.backend.api.auth.dto.UserDto;
-import gitbal.backend.domain.refreshtoken.RefreshTokenService;
+import gitbal.backend.domain.refreshtoken.application.RefreshTokenService;
 import gitbal.backend.domain.user.User;
 import gitbal.backend.domain.user.UserRepository;
 import gitbal.backend.domain.majorlanguage.application.MajorLanguageService;
