+++ src/main/java/gitbal/backend/domain/refreshtoken/application/repository/RefreshTokenRepository.java
@@ -1,13 +1,14 @@
-package gitbal.backend.domain.refreshtoken;
-
+package gitbal.backend.domain.refreshtoken.application.repository;
 
 import gitbal.backend.domain.refreshtoken.RefreshToken;
 import java.util.Optional;
-import org.springframework.data.repository.CrudRepository;
 
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
+public interface RefreshTokenRepository {
     Optional<RefreshToken> findByUserNickname(String userNickname);
 
     void deleteByUserNickname(String userNickname);
 
+    void deleteById(String id);
+
+    void save(RefreshToken refreshToken);
 }
