+++ src/main/java/gitbal/backend/global/security/jwt/JwtTokenProvider.java
@@ -2,10 +2,10 @@ package gitbal.backend.global.security.jwt;
 
 
 import gitbal.backend.domain.refreshtoken.RefreshToken;
+import gitbal.backend.domain.refreshtoken.application.repository.RefreshTokenRepository;
 import gitbal.backend.domain.user.User;
 import gitbal.backend.global.security.CustomUserDetails;
 import gitbal.backend.global.security.GithubOAuth2UserInfo;
-import gitbal.backend.domain.refreshtoken.RefreshTokenRepository;
 import gitbal.backend.domain.user.UserRepository;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
